Better way to round grades
When we scale the grading and have a cutoff at which it is not passed,
we have previously used the MBF computation which has a linear scaling
factor of 3 for university grades. This means that even grades with are
slightly below the cutoff line may pass after rounding. We previously
had a special condition for this case. However, it is better to modify
the coefficient to 3.05, which guarantees a grade of 4.05. Annoying
while this works in Python (round to even) where round(4.05) = 4.0, it
does not in Excel where ROUND(4.05) = 4.1. So we need to modify the
Excel export to find a general solution that is consistent.